rename sempahore utils
